+++ bank-feeds/src/codat_bankfeeds/models/errors/__init__.py
@@ -3,4 +3,5 @@
 from .errormessage import ErrorMessage, ErrorMessageData
 from .sdkerror import SDKError
 
+
 __all__ = ["ErrorMessage", "ErrorMessageData", "SDKError"]